에러 처리 수정
